+++ src/utils/codeframe.ts
@@ -4,6 +4,12 @@ import { type Result } from "../result";
 
 const prefix = ["", "WARNING", "ERROR"];
 
+const supportedFilenames = ["package.json", "package-lock.json"];
+
+function supportsCodeFrame(filePath: string): boolean {
+	return supportedFilenames.some((name) => filePath.endsWith(name));
+}
+
 function formatMessage(
 	content: string,
 	result: Pick<Result, "filePath">,
@@ -12,7 +18,7 @@ function formatMessage(
 	const preamble = `${prefix[message.severity]}: ${message.message} (${message.ruleId}) at ${
 		result.filePath
 	}`;
-	const formatted = result.filePath.endsWith("package.json")
+	const formatted = supportsCodeFrame(result.filePath)
 		? codeFrameColumns(content, { start: message, end: message }, { highlightCode: false })
 		: "";
 	return [preamble, formatted].join("\n");
@@ -28,8 +34,7 @@ function isResultArray(value: Result[] | Generator<Message>): value is Result[] 
 export function codeframe(content: string, errors: Result[] | Generator<Message>): string {
 	if (isResultArray(errors)) {
 		return errors
-			.map((it) => it.messages.map((jt) => formatMessage(content, it, jt)))
-			.flat()
+			.flatMap((it) => it.messages.map((jt) => formatMessage(content, it, jt)))
 			.join("\n\n");
 	} else {
 		const messages = Array.from(errors);